Add ABI export decoration
Use CMake's GenerateExportHeader to generate a header with an
appropriate definition of an ABI export symbol. Add such annotation to
components which need it. Don't export symbols by default on ELF
platforms.

This allows building of shared libraries on Windows to work properly (as
these require ABI export/import decoration).

Note that this bumps the minimum CMake requirement from 2.6 to 2.8.12.